+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.4.21] - 2023-08-13
+- Add `Inventory.to_csv()` and `InventoryHP.to_csv()` methods for writing an inventory's contents
+to a CSV file (in user's chosen units) (#94).
+
 ## [0.4.20] - 2023-08-08
 - Add `rd.read_csv()` function for creating an inventory by reading nuclides & quantities (and
 optionally units) in from a CSV file (#94).